refactor: move port code to notated place
